+++ src/test/unit/System.Windows.Forms/System/Windows/Forms/CommonDialogTests.cs
@@ -5,7 +5,6 @@
 
 using System.ComponentModel;
 using System.Reflection;
-using Microsoft.DotNet.RemoteExecutor;
 using Moq;
 
 namespace System.Windows.Forms.Tests;
@@ -118,30 +117,10 @@ public class CommonDialogTests
         Assert.Equal(expectedDialogResult, dialog.ShowDialog(owner.Object));
     }
 
-    [WinFormsTheory(Skip = "Crash with AbandonedMutexException. See: https://github.com/dotnet/arcade/issues/5325")]
-    [InlineData(true, DialogResult.OK)]
-    [InlineData(false, DialogResult.Cancel)]
-    public void ShowDialog_NonControlOwnerWithVisualStyles_ReturnsExpected(bool runDialogResultParam, DialogResult expectedDialogResultParam)
-    {
-        // Run this from another thread as we call Application.EnableVisualStyles.
-        RemoteExecutor.Invoke((runDialogResultString, expectedDialogResultString) =>
-        {
-            bool runDialogResult = bool.Parse(runDialogResultString);
-            DialogResult expectedDialogResult = (DialogResult)Enum.Parse(typeof(DialogResult), expectedDialogResultString);
-
-            Application.EnableVisualStyles();
-
-            using SubCommonDialog dialog = new()
-            {
-                RunDialogResult = runDialogResult
-            };
-            var owner = new Mock<IWin32Window>(MockBehavior.Strict);
-            owner
-                .Setup(o => o.Handle)
-                .Returns(IntPtr.Zero);
-            Assert.Equal(expectedDialogResult, dialog.ShowDialog(owner.Object));
-        }, runDialogResultParam.ToString(), expectedDialogResultParam.ToString()).Dispose();
-    }
+    // ShowDialog_NonControlOwnerWithVisualStyles_ReturnsExpected,
+    // ShowDialog_ControlOwnerWithVisualStyles_ReturnsExpected,
+    // ShowDialog_ControlOwnerWithHandleWithVisualStyles_ReturnsExpected moved to
+    // src/test/integration/UIIntegrationTests/CommonDialogTests.cs (issue #4500).
 
     [WinFormsTheory]
     [InlineData(true, DialogResult.OK)]
@@ -156,28 +135,6 @@ public class CommonDialogTests
         Assert.Equal(expectedDialogResult, dialog.ShowDialog(owner));
     }
 
-    [WinFormsTheory(Skip = "Crash with AbandonedMutexException. See: https://github.com/dotnet/arcade/issues/5325")]
-    [InlineData(true, DialogResult.OK)]
-    [InlineData(false, DialogResult.Cancel)]
-    public void ShowDialog_ControlOwnerWithVisualStyles_ReturnsExpected(bool runDialogResultParam, DialogResult expectedDialogResultParam)
-    {
-        // Run this from another thread as we call Application.EnableVisualStyles.
-        RemoteExecutor.Invoke((runDialogResultString, expectedDialogResultString) =>
-        {
-            bool runDialogResult = bool.Parse(runDialogResultString);
-            DialogResult expectedDialogResult = (DialogResult)Enum.Parse(typeof(DialogResult), expectedDialogResultString);
-
-            Application.EnableVisualStyles();
-
-            using SubCommonDialog dialog = new()
-            {
-                RunDialogResult = runDialogResult
-            };
-            using Control owner = new();
-            Assert.Equal(expectedDialogResult, dialog.ShowDialog(owner));
-        }, runDialogResultParam.ToString(), expectedDialogResultParam.ToString()).Dispose();
-    }
-
     [WinFormsTheory]
     [InlineData(true, DialogResult.OK)]
     [InlineData(false, DialogResult.Cancel)]
@@ -190,29 +147,6 @@ public class CommonDialogTests
         using Control owner = new();
         Assert.NotEqual(IntPtr.Zero, owner.Handle);
         Assert.Equal(expectedDialogResult, dialog.ShowDialog(owner));
-    }
-
-    [WinFormsTheory(Skip = "Crash with AbandonedMutexException. See: https://github.com/dotnet/arcade/issues/5325")]
-    [InlineData(true, DialogResult.OK)]
-    [InlineData(false, DialogResult.Cancel)]
-    public void ShowDialog_ControlOwnerWithHandleWithVisualStyles_ReturnsExpected(bool runDialogResultParam, DialogResult expectedDialogResultParam)
-    {
-        // Run this from another thread as we call Application.EnableVisualStyles.
-        RemoteExecutor.Invoke((runDialogResultString, expectedDialogResultString) =>
-        {
-            bool runDialogResult = bool.Parse(runDialogResultString);
-            DialogResult expectedDialogResult = (DialogResult)Enum.Parse(typeof(DialogResult), expectedDialogResultString);
-
-            Application.EnableVisualStyles();
-
-            using SubCommonDialog dialog = new()
-            {
-                RunDialogResult = runDialogResult
-            };
-            using Control owner = new();
-            Assert.NotEqual(IntPtr.Zero, owner.Handle);
-            Assert.Equal(expectedDialogResult, dialog.ShowDialog(owner));
-        }, runDialogResultParam.ToString(), expectedDialogResultParam.ToString()).Dispose();
     }
 
     [WinFormsFact]
